cแก้เป็น Float

+++ Lab2_1.c
@@ -2,10 +2,10 @@
 
 void main()
 {
-    int a,b,c;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    scanf("%d",&c);
+    float a,b,c;
+    scanf("%f",&a);
+    scanf("%f",&b);
+    scanf("%f",&c);
 
     int tt=a+b+c;
     if(tt >= 80)
